+++ embedded_scheme/scheme_proc.h
@@ -1,4 +1,4 @@
-extern LANGSPEC bool SCHEME_mousepress(int button, int x, int y);
-extern LANGSPEC bool SCHEME_mousemove(int button, int x, int y);
-extern LANGSPEC bool SCHEME_mouserelease(int button, int x, int y);
+extern LANGSPEC bool SCHEME_mousepress(int button, float x, float y);
+extern LANGSPEC bool SCHEME_mousemove(int button, float x, float y);
+extern LANGSPEC bool SCHEME_mouserelease(int button, float x, float y);
 extern LANGSPEC void SCHEME_start(void);
